zip: read zip64 records so members past 2 GiB open
A 3.36 GB benchmark archive failed on 300 of its members with "corrupt
local file header": Python's zipfile writes zip64 records once an entry
offset passes 2 GiB, and ZipReader read the 0xFFFFFFFF sentinels as
literal offsets.

Parse the zip64 EOCD (via its locator) for the directory-level fields and
the 0x0001 extra field per entry, and widen Entry's size/offset fields to
64 bits.

The fixture is written with force_zip64 so it carries the same records at
a few hundred bytes.

+++ src/zip_reader.h
@@ -19,11 +19,13 @@ class ZipReader {
 public:
     struct Entry {
         std::string name;               // UTF-8 (CP949 legacy names converted)
-        uint32_t compressed_size = 0;
-        uint32_t uncompressed_size = 0;
+        // 64-bit: zip64 archives (>4GiB, or offsets past 2GiB from writers
+        // like Python's zipfile) store these in the zip64 extra field.
+        uint64_t compressed_size = 0;
+        uint64_t uncompressed_size = 0;
         uint16_t method = 0;            // 0=STORE, 8=DEFLATE
         uint16_t flags = 0;             // general purpose bit flags
-        uint32_t local_header_offset = 0;
+        uint64_t local_header_offset = 0;
     };
 
     explicit ZipReader(const std::string& path);
